requirements.txt
astroid==2.2.5
boto3==1.9.212
botocore==1.12.212
certifi==2019.6.16
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.15.2
gunicorn==19.9.0
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wincertstore==0.2
wrapt==1.11.2
